# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Heymaty - OpenAPI 3.0",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://heymaty.com/"
    }
  ],
...
  "paths": {
...
    "/comments/post/{id}": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "get a comment",
        "description": "get a comment",
        "operationId": "getComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "object",
                      "properties": {
                        "docs": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Comments"
                          }
                        },
                        "totalDocs": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "pagingCounter": {
                          "type": "number"
                        },
                        "hasPrevPage": {
                          "type": "boolean"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "prevPage": {
                          "type": "number"
                        },
                        "nextPage": {
                          "type": "number"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/comments/count/post/{id}": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "get a count comments",
        "description": "get a count comments",
        "operationId": "getCountComments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
    "/comments/": {
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "create a comment",
        "description": "create a comment",
        "operationId": "createComment",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "postId": {
                    "type": "string"
                  },
                  "comment": {
                    "type": "string"
                  },
                  "mentions": {
                    "type": "integer"
                  }
                },
                "required": [
                  "postId",
                  "comment"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "newComment": {
                      "type": "object",
                      "$ref": "#/components/schemas/Comments"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
    "/comments/{id}": {
      "delete": {
        "tags": [
          "comments"
        ],
        "summary": "delete a comment",
        "description": "delete a comment",
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Service": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the service."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the service was created.",
            "default": "The current date-time"
          }
        }
      },
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "msg": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "errors"
        ]
      },
...
      "Comments": {
        "required": [
          "postId",
          "userId",
          "comment"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "userId": {
            "$ref": "#/components/schemas/User"
          },
          "postId": {
            "$ref": "#/components/schemas/posts"
          },
          "comment": {
            "type": "string",
            "maxLength": 200
          },
          "mentions": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 30,
            "description": "Name of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user"
          },
          "emailConfirmed": {
            "type": "boolean",
            "default": false
          },
          "emailConfirmationCode": {
            "type": "string"
          },
          "emailConfirmationValidUntil": {
            "type": "string",
            "format": "date-time"
          },
          "emailConfirmationRequestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "girl",
              "admin",
              "moderator"
            ],
            "default": "user"
          },
          "verification": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "subscribers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscribed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscriptionEnabled": {
            "type": "boolean",
            "default": false
          },
          "profileNum": {
            "type": "integer",
            "format": "int64"
          },
          "profileUrl": {
            "type": "string"
          },
          "profileGallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "verified": {
            "type": "boolean",
            "default": false
          },
          "balance": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "unconfirmedBalance": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "inBTCAddress": {
            "type": "string"
          },
          "outBTCAddress": {
            "type": "string"
          },
          "paymentCode": {
            "type": "string"
          },
          "invoice": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "default": 18
          },
          "phone": {
            "type": "string"
          },
          "country": {
            "type": "object"
          },
          "state": {
            "type": "object"
          },
          "city": {
            "type": "object"
          },
          "profile": {
            "type": "object"
          },
          "cover": {
            "type": "object"
          },
          "nationalId": {
            "type": "object"
          },
          "pricePerDay": {
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "pricePerMessage": {
            "type": "integer",
            "format": "int32",
            "default": 9999
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            }
          },
          "urlTelegram": {
            "type": "string",
            "format": "uri"
          },
          "urlWhatsApp": {
            "type": "string",
            "format": "uri"
          },
          "loginAttempts": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "blockExpires": {
            "type": "string",
            "format": "date-time",
            "default": "2023-09-14T00:00:00Z"
          },
          "googleId": {
            "type": "string"
          },
          "messageBundles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": {
                  "$ref": "#/components/schemas/User"
                },
                "messages": {
                  "type": "number"
                }
              }
            }
          },
          "personalFirstName": {
            "type": "string"
          },
          "personalSecondName": {
            "type": "string"
          },
          "personalFamilyName": {
            "type": "string"
          },
          "personalDayOfBirth": {
            "type": "integer",
            "format": "int32"
          },
          "personalMonthOfBirth": {
            "type": "integer",
            "format": "int32"
          },
          "personalYearOfBirth": {
            "type": "integer",
            "format": "int32"
          },
          "personalSex": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ]
          },
          "personalCountry": {
            "type": "string"
          },
          "personalCity": {
            "type": "string"
          },
          "personalStreet": {
            "type": "string"
          },
          "personalBuilding": {
            "type": "string"
          },
          "personalAppartment": {
            "type": "string"
          },
          "personalZIP": {
            "type": "string"
          },
          "personalPassport": {
            "type": "object"
          },
          "personalIDCardFront": {
            "type": "object"
          },
          "personalIDCardBack": {
            "type": "object"
          },
          "personalDriverLicenseFront": {
            "type": "object"
          },
          "personalDriverLicenseBack": {
            "type": "object"
          },
          "personalIDExpirationDay": {
            "type": "integer",
            "format": "int32"
          },
          "personalIDExpirationMonth": {
            "type": "integer",
            "format": "int32"
          },
          "personalIDExpirationYear": {
            "type": "integer",
            "format": "int32"
          },
          "personalIDHasNoExpiration": {
            "type": "boolean",
            "default": false
          },
          "personalSelfie": {
            "type": "object"
          }
        }
      },
      "Language": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the service."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "2023-09-14T00:00:00Z"
          }
        }
      },
      "posts": {
        "required": [
          "title",
          "description"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "userId": {
            "$ref": "#/components/schemas/User"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "videos": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}